requirements.txt
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
debugpy==1.8.19
docker==7.1.0
flake8==7.3.0
idna==3.11
mccabe==0.7.0
packaging==25.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
pyhumps==3.8.0
PyYAML==6.0.3
requests==2.32.5
termcolor==3.2.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.2
websocket-client==1.9.0
